refacto: Move functions definitions into a separate file.
